Refactor chktex files to allow for more linters
